chore: rename panel to control-panel
